Add tests for multiple enums in one file and missing enums
Add the env variable GOLINE=..., needed to vet the relevant const/var block.

Add a test case with multiple enum definition in the same file.

Add a test case to expect an error when the const/var block is not found in
the line following the `//go:generate ...` comment.

Remove unneeded output files from internal/testdata (uneeded copies of the
same files in internal/testdata/exp).